# openapi/api.yaml
...
openapi: 3.0.3
info:
  title: SurenSaaS API
  version: 1.0.0
  description: API multi-PME pour construction, nettoyage et diagnostic énergétique
servers:
  - url: https://back-xxx.run.app/api/v1
    description: Production
paths:
...
  /admin/telegram/bots:
    get:
      summary: Lister les bots Telegram disponibles
      security:
        - bearerAuth: []
      description: Nécessite le rôle admin. Retourne tous les bots configurés pour l'environnement actuel.
      responses:
        200:
          description: Liste des bots disponibles
          content:
            application/json:
              schema:
                type: object
                properties:
                  bots:
                    type: array
                    items:
                      $ref: '#/components/schemas/TelegramBotInfo'
                  environment:
                    type: string
                    enum: [test, prod]
        403:
          description: Admin requis
  /admin/users/{user_id}/telegram-invitation:
    post:
      summary: Générer un lien d'invitation Telegram
      security:
        - bearerAuth: []
      description: |
        Génère un lien d'invitation pour un utilisateur vers un bot Telegram spécifique.
        Le lien expire après 7 jours par défaut.
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelegramInvitationRequest'
      responses:
        200:
          description: Lien d'invitation généré
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TelegramInvitationResponse'
        400:
          description: Paramètres invalides ou bot non configuré
        403:
          description: Admin requis
        404:
          description: Utilisateur non trouvé

  # ============================================
  # TELEGRAM WEBHOOK (PUBLIC)
  # ============================================
...
  /admin/pre-authorized-emails:
    get:
      summary: Lister les emails pré-autorisés
      security:
        - bearerAuth: []
      description: Nécessite le rôle admin
      responses:
        200:
          description: Liste des emails pré-autorisés
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PreAuthorizedEmail'
        403:
          description: Admin requis

    post:
      summary: Ajouter un email pré-autorisé
      security:
        - bearerAuth: []
      description: Nécessite le rôle admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreAuthorizedEmailCreate'
      responses:
        201:
          description: Email pré-autorisé créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreAuthorizedEmail'
        400:
          description: Données invalides
        403:
          description: Admin requis
        409:
          description: Email déjà pré-autorisé
  /admin/pre-authorized-emails/{email_id}:
    patch:
      summary: Modifier un email pré-autorisé
      security:
        - bearerAuth: []
      description: Nécessite le rôle admin
      parameters:
        - name: email_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreAuthorizedEmailUpdate'
      responses:
        200:
          description: Email pré-autorisé mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreAuthorizedEmail'
        403:
          description: Admin requis
        404:
          description: Email pré-autorisé non trouvé

    delete:
      summary: Supprimer un email pré-autorisé
      security:
        - bearerAuth: []
      description: Nécessite le rôle admin
      parameters:
        - name: email_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: Email pré-autorisé supprimé
        403:
          description: Admin requis
        404:
          description: Email pré-autorisé non trouvé

  # ============================================
  # ADMIN - USER CAPABILITIES
  # ============================================
  /admin/users/{user_id}/capabilities:
    get:
      summary: Lister les capabilities d'un utilisateur
      security:
        - bearerAuth: []
      description: Nécessite le rôle admin
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Liste des capabilities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserCapability'
        403:
          description: Admin requis

    post:
      summary: Assigner une capability à un utilisateur
      security:
        - bearerAuth: []
      description: Nécessite le rôle admin
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCapabilityCreate'
      responses:
        201:
          description: Capability assignée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCapability'
        400:
          description: Données invalides
        403:
          description: Admin requis
        404:
          description: Utilisateur non trouvé
  /admin/users/{user_id}/capabilities/{capability_id}:
    put:
      summary: Modifier une capability d'un utilisateur
      security:
        - bearerAuth: []
      description: Nécessite le rôle admin
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: capability_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCapabilityUpdate'
      responses:
        200:
          description: Capability mise à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCapability'
        403:
          description: Admin requis
        404:
          description: Capability non trouvée

    delete:
      summary: Retirer une capability d'un utilisateur
      security:
        - bearerAuth: []
      description: Nécessite le rôle admin
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: capability_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: Capability retirée
        403:
          description: Admin requis
        404:
          description: Capability non trouvée

  # ============================================
  # ADMIN - ORGANIZATION CAPABILITIES
  # ============================================
  /admin/organization-capabilities:
    get:
      summary: Lister les capabilities de l'organisation
      security:
        - bearerAuth: []
      description: Nécessite le rôle admin
      responses:
        200:
          description: Liste des capabilities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrganizationCapability'
        403:
          description: Admin requis

    post:
      summary: Créer une capability pour l'organisation
      security:
        - bearerAuth: []
      description: Nécessite le rôle admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationCapabilityCreate'
      responses:
        201:
          description: Capability créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationCapability'
        400:
          description: Données invalides
        403:
          description: Admin requis
        409:
          description: Capability existe déjà
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PreAuthorizedEmail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        org_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, user]
          default: user
        is_active:
          type: boolean
          default: true
        expires_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PreAuthorizedEmailCreate:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, user]
          default: user
        expires_at:
          type: string
          format: date-time
          nullable: true
    PreAuthorizedEmailUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [admin, user]
        is_active:
          type: boolean
        expires_at:
          type: string
          format: date-time
          nullable: true
    OrganizationCapability:
      type: object
      properties:
        id:
          type: string
          format: uuid
        org_id:
          type: string
          format: uuid
        capability_name:
          type: string
          example: "construction:facturation:read"
        description:
          type: string
        is_active:
          type: boolean
          default: true
        created_at:
          type: string
          format: date-time
    OrganizationCapabilityCreate:
      type: object
      required:
        - capability_name
      properties:
        capability_name:
          type: string
          example: "construction:facturation:read"
        description:
          type: string
        is_active:
          type: boolean
          default: true
    UserCapability:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        capability_name:
          type: string
          example: "construction:facturation:read"
        resource_type:
          type: string
          enum: [company, global]
          default: global
        resource_id:
          type: string
          format: uuid
          nullable: true
        granted_by:
          type: string
          format: uuid
        granted_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
          nullable: true
        is_active:
          type: boolean
          default: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserCapabilityCreate:
      type: object
      required:
        - capability_name
      properties:
        capability_name:
          type: string
          example: "construction:facturation:read"
        resource_type:
          type: string
          enum: [company, global]
          default: global
        resource_id:
          type: string
          format: uuid
          nullable: true
        expires_at:
          type: string
          format: date-time
          nullable: true
    UserCapabilityUpdate:
      type: object
      properties:
        resource_type:
          type: string
          enum: [company, global]
        resource_id:
          type: string
          format: uuid
          nullable: true
        is_active:
          type: boolean
        expires_at:
          type: string
          format: date-time
          nullable: true

    # Telegram Bot Schemas
    TelegramBotInfo:
      type: object
      properties:
        bot_id:
          type: string
          description: Identifiant unique du bot (e.g., 'construction', 'audit')
          example: "construction"
        bot_name:
          type: string
          description: Nom affiché du bot
          example: "Construction"
        description:
          type: string
          description: Description des fonctionnalités
          example: "Envoi et traitement des factures de construction"
        icon:
          type: string
          description: Emoji pour l'interface
          example: "🏗️"
        username:
          type: string
          description: Username Telegram du bot
          example: "suren_construction_bot"
        configured:
          type: boolean
          description: Indique si le bot est configuré
    TelegramInvitationRequest:
      type: object
      required:
        - bot_id
      properties:
        bot_id:
          type: string
          description: ID du bot vers lequel inviter l'utilisateur
          example: "construction"
        expires_in_hours:
          type: integer
          description: Durée de validité du lien en heures
          default: 168
          minimum: 1
          maximum: 720
    TelegramInvitationResponse:
      type: object
      properties:
        telegram_link:
          type: string
          description: Lien d'invitation complet
          example: "https://t.me/suren_construction_bot?start=ABC123..."
        bot_id:
          type: string
          description: ID du bot
          example: "construction"
        bot_name:
          type: string
          description: Nom du bot
          example: "Construction"
        bot_username:
          type: string
          description: Username Telegram
          example: "suren_construction_bot"
        bot_icon:
          type: string
          description: Emoji du bot
          example: "🏗️"
        description:
          type: string
          description: Description du bot
          example: "Envoi et traitement des factures de construction"
        expires_at:
          type: string
          format: date-time
          description: Date d'expiration du lien
        user_id:
          type: string
          format: uuid
          description: ID de l'utilisateur invité
        org_id:
          type: string
          format: uuid
          description: ID de l'organisation